Move project extension
