+++ snuba/query/extensions.py
@@ -1,13 +1,7 @@
 from abc import ABC
 
-from snuba.query.query_processor import (
-    DummyExtensionProcessor,
-    ExtensionQueryProcessor,
-    ProjectExtensionProcessor
-)
-from snuba.schemas import (
-    Schema
-)
+from snuba.query.query_processor import ExtensionQueryProcessor
+from snuba.schemas import Schema
 
 
 class QueryExtension(ABC):
@@ -29,31 +23,3 @@ class QueryExtension(ABC):
 
     def get_processor(self) -> ExtensionQueryProcessor:
         return self.__processor
-
-
-PROJECT_EXTENSION_SCHEMA = {
-    'type': 'object',
-    'properties': {
-        'project': {
-            'anyOf': [
-                {'type': 'integer', 'minimum': 1},
-                {
-                    'type': 'array',
-                    'items': {'type': 'integer', 'minimum': 1},
-                    'minItems': 1,
-                },
-            ]
-        },
-    },
-    # Need to select down to the project level for customer isolation and performance
-    'required': ['project'],
-    'additionalProperties': False,
-}
-
-
-class ProjectExtension(QueryExtension):
-    def __init__(self) -> None:
-        super().__init__(
-            schema=PROJECT_EXTENSION_SCHEMA,
-            processor=ProjectExtensionProcessor(),
-        )
